[K/N] Add tests to verify lldb linenum bug
When calling an inlined function in another file, LLDB sometimes steps
to meaningless lines. This commit adds a test that verifies that bug.

The test does not impose any specific stepping order, since this would
be too fragile. It only tests that LLDB does not step to empty (or fully
//-commented) lines.